requirement.txt
langchain>=0.1.0
langchain-community
langchain-google-genai
langchain-chroma
chromadb>=0.4.22
pydantic>=2.0.0
sentence-transformers
streamlit
python-dotenv
pypdf
docx2txt
google-generativeai